fixes for views for generic array metafunctions

+++ include/nmtools/array/view/matmul.hpp
@@ -45,8 +45,8 @@ namespace nmtools::view::detail
         // auto r1 = at(bshape,i1);
         auto r2 = at(bshape,i2);
 
-        auto adim = len(ashape);
-        auto bdim = len(bshape);
+        auto adim = (size_t)len(ashape);
+        auto bdim = (size_t)len(bshape);
         const auto [h_dim, l_dim] = (adim > bdim) ? nmtools_tuple{adim,bdim} : nmtools_tuple{bdim,adim};
 
         // check matrix shape
